+++ R/textscale-package.R
@@ -3,6 +3,7 @@
 
 ## usethis namespace: start
 #' @importFrom dplyr mutate select filter
+#' @importFrom stats coef plogis predict qnorm quantile
 #' @importFrom tibble tibble
 ## usethis namespace: end
 NULL